Added initial tests and changed code accordingly.
Using Supertest and Mocha to run the tests. Rewrote the server.js to
export the Koa app so that test runner could initialize it. Made sure
that logging is disabled while test are tunning (environment as test),
this should keep the output nice and readable. A few other small changes
were also made, all for the sake of running tests. Running tests could
be actived by typing: npm test