Resolves #16, added error handler based on linked example.
The error handler will try to make sure that API endpoints always
respond with JSON, evene if errors are thrown. It will also allow the
message to be sent within the response in case the status code was
provided (this should indicate an intention), otherwise, the negeral
code 500 and JSON payload with statdard message would be used. This
implementation could change in future.