Fixes #15, added better flow that should work as expected.
Also added a bit different logic for sending error responses,
this.throw() is being used. There is an error handling middleware near
the top of the middleware chain. It will catch any errors and respond
with corresponding code and ensure the response if formatted as JSON.
All the CORS filters are also running because we are dealing with a
normal response flow here.