References #37, backend side implementation, unfinished.
Initial implementation is in. Works fairly well (with some limitations)
when used purely on the backend side. A connecot methods for the
UI/client side are still unfinished. Those should be used to finish the
flow. The code is still very unclean (even dirty) and has to be modified
to allow proper error handling and possibility to use calles from the
generators (proper error handling and such things).