References #50, added initial socket implementation.
The socket implementation is incomplete and should be extended. It does
have the initial event listeners and authentication checkes in place. It
alo has an apiAction event listener that would later be used to catch
the events emitted by API actions and send data through sockets to any
clients listening (used in single project view).