References #61 and #63, changed download logic.
Changed value for content-disposition to inline. This would allow the
browser to show the file if possible and fall back to just downloading
if not. File name in case of trying to save the inline file should be
the same as original, at least in fresh versions of browsers (and we do
not support any of the archaic ones).