References #68, added loading in parts possibility.
Messages and activities are loaded in batched by default. It is possible
to provide the maximum size of the batch, although 50 is set as a
maximum allowed limit. It is also possible to provide the creation date
for the oldest loaded object, this will be used as a starting point to
load even older ones.