References #82, added activity stream and socket.
Introduces own events for action cases. Extended WebSocket and
ActivityStream logic to handle those. Socket defaults to sending out the
task:update event as the front-end will handle it like that for now
(might change to be more specific in the future). Extended activity
stream to handle the more advanced logic needed by multiple objects
being part of the action (participant and task, resource and task,
outcome and task). The logic got extended, not rewritten. This should
keep the compatibility with previous activities (already stored). In
case of some more drastic changes, the logic should probably be
rewritten in full, keeping it compatible with previous version is not
required (as no public instances exist, yet).